+++ php/Terminus/Commands/OrganizationsCommand.php
@@ -24,8 +24,8 @@ class OrganizationsCommand extends TerminusCommand {
    * @return OrganizationsCommand
    */
   public function __construct(array $options = []) {
+    $options['require_login'] = true;
     parent::__construct($options);
-    $this->helpers->auth->ensureLogin();
     $this->sites = new Sites();
   }