+++ php/utils.php
@@ -3,7 +3,6 @@
 namespace Terminus\Utils;
 
 use ArrayIterator;
-use Symfony\Component\Yaml\Yaml;
 use Terminus\Caches\FileCache;
 use Terminus\Commands\TerminusCommand;
 use Terminus\Dispatcher;
@@ -70,32 +69,6 @@ function checkForUpdate($logger) {
 }
 
 /**
- * Sets constants necessary for the proper functioning of Terminus
- *
- * @return void
- */
-function defineConstants() {
-  if (!defined('Terminus')) {
-    define('Terminus', true);
-  }
-  $default_constants = Yaml::parse(
-    file_get_contents(TERMINUS_ROOT . '/config/constants.yml')
-  );
-  foreach ($default_constants as $var_name => $default) {
-    if (isset($_SERVER[$var_name]) && ($_SERVER[$var_name] != '')) {
-      define($var_name, $_SERVER[$var_name]);
-    } else if (!defined($var_name)) {
-      define($var_name, $default);
-    }
-  }
-  date_default_timezone_set(TERMINUS_TIME_ZONE);
-
-  if (!defined('TERMINUS_SCRIPT')) {
-    define('TERMINUS_SCRIPT', 'php/Terminus.php');
-  }
-}
-
-/**
  * Ensures that the given destination is valid
  *
  * @param string $destination Location of directory to ensure viability of
@@ -143,19 +116,6 @@ function getVendorPaths() {
     TERMINUS_ROOT . '/vendor'
   ];
   return $vendor_paths;
-}
-
-/**
- * Imports environment variables
- *
- * @return void
- */
-function importEnvironmentVariables() {
-  //Load environment variables from __DIR__/.env
-  if (file_exists(getcwd() . '/.env')) {
-    $env = new \Dotenv\Dotenv(getcwd());
-    $env->load();
-  }
 }
 
 /**